+++ ChangeLog.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - A target layer can be defined for instantiated GameObjects via `GameObjectInstantiator.Settings.layer` (thanks [Krzysztof Lesiak][Holo-Krzysztof] for #393)
+- Re-normalize bone weights (always for design-time import and opt-in at runtime via `GLTFAST_SAFE` scripting define)
 ### Changed
 - Mecanim (non-legacy) is now the default for importing animation clips at design-time (thanks [@hybridherbst][hybridherbst] for #388)
+- All four bone weights are imported at design-time, regardless of quality setting
 ### Fixed
 - Fail more gracefully when parsing invalid JSON
 - Proper error handling on glTF-binary files with invalid chunks (unknown type or invalid length; #389)